build: Change application's package name
Cause: Unable to add OAuth client to Google API due to duplicated
package name and SHA1 fingerprint
