--- app/src/main/java/com/example/stren/app/MainActivity.kt
@@ -1,38 +1,0 @@
-package com.example.stren.app
-
-import android.os.Bundle
-import androidx.activity.ComponentActivity
-import androidx.activity.compose.LocalActivityResultRegistryOwner
-import androidx.activity.compose.setContent
-import androidx.compose.foundation.layout.fillMaxSize
-import androidx.compose.runtime.CompositionLocalProvider
-import androidx.compose.runtime.staticCompositionLocalOf
-import androidx.compose.ui.Modifier
-import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
-import com.example.stren.ui.theme.StrenTheme
-import com.facebook.CallbackManager
-import dagger.hilt.android.AndroidEntryPoint
-
-val LocalFacebookCallbackManager =
-    staticCompositionLocalOf<CallbackManager> { error("No CallbackManager provided") }
-
-@AndroidEntryPoint
-class MainActivity : ComponentActivity() {
-    private var facebookCallbackManager = CallbackManager.Factory.create();
-
-    override fun onCreate(savedInstanceState: Bundle?) {
-
-        installSplashScreen()
-        super.onCreate(savedInstanceState)
-        setContent {
-            StrenTheme {
-                CompositionLocalProvider(
-                    LocalFacebookCallbackManager provides facebookCallbackManager,
-                    LocalActivityResultRegistryOwner provides this
-                ) {
-                    StrenApp(modifier = Modifier.fillMaxSize())
-                }
-            }
-        }
-    }
-}
