--- app/src/main/java/com/example/stren/app/StrenApplication.kt
@@ -1,7 +1,0 @@
-package com.example.stren.app
-
-import android.app.Application
-import dagger.hilt.android.HiltAndroidApp
-
-@HiltAndroidApp
-class StrenApplication : Application()
