--- app/src/main/java/com/example/stren/core/service/di/ServiceModule.kt
@@ -1,15 +1,0 @@
-package com.example.stren.core.service.di
-
-import com.example.stren.core.service.AuthenticationService
-import com.example.stren.core.service.AuthenticationServiceImpl
-import dagger.Binds
-import dagger.Module
-import dagger.hilt.InstallIn
-import dagger.hilt.android.components.ViewModelComponent
-
-@Module
-@InstallIn(ViewModelComponent::class)
-abstract class ServiceModule {
-    @Binds
-    abstract fun provideAuthService(impl: AuthenticationServiceImpl): AuthenticationService
-}
