+++ app/src/main/java/com/haidoan/android/stren/app/MainActivity.kt
@@ -1,0 +1,38 @@
+package com.haidoan.android.stren.app
+
+import android.os.Bundle
+import androidx.activity.ComponentActivity
+import androidx.activity.compose.LocalActivityResultRegistryOwner
+import androidx.activity.compose.setContent
+import androidx.compose.foundation.layout.fillMaxSize
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.staticCompositionLocalOf
+import androidx.compose.ui.Modifier
+import androidx.core.splashscreen.SplashScreen.Companion.installSplashScreen
+import com.facebook.CallbackManager
+import com.haidoan.android.stren.designsystem.theme.StrenTheme
+import dagger.hilt.android.AndroidEntryPoint
+
+val LocalFacebookCallbackManager =
+    staticCompositionLocalOf<CallbackManager> { error("No CallbackManager provided") }
+
+@AndroidEntryPoint
+class MainActivity : ComponentActivity() {
+    private var facebookCallbackManager = CallbackManager.Factory.create()
+
+    override fun onCreate(savedInstanceState: Bundle?) {
+
+        installSplashScreen()
+        super.onCreate(savedInstanceState)
+        setContent {
+            StrenTheme {
+                CompositionLocalProvider(
+                    LocalFacebookCallbackManager provides facebookCallbackManager,
+                    LocalActivityResultRegistryOwner provides this
+                ) {
+                    StrenApp(modifier = Modifier.fillMaxSize())
+                }
+            }
+        }
+    }
+}
