+++ app/src/main/java/com/haidoan/android/stren/app/StrenApplication.kt
@@ -1,0 +1,7 @@
+package com.haidoan.android.stren.app
+
+import android.app.Application
+import dagger.hilt.android.HiltAndroidApp
+
+@HiltAndroidApp
+class StrenApplication : Application()
