+++ app/src/main/java/com/haidoan/android/stren/core/service/AuthenticationService.kt
@@ -1,0 +1,11 @@
+package com.haidoan.android.stren.core.service
+
+import com.facebook.AccessToken
+
+interface AuthenticationService {
+    fun addAuthStateListeners(onUserAuthenticated: () -> Unit, onUserNotAuthenticated: () -> Unit)
+    suspend fun authenticate(email: String, password: String)
+
+    suspend fun signUp(email: String, password: String)
+    suspend fun authenticateWithFacebook(token: AccessToken)
+}
