+++ app/src/main/java/com/haidoan/android/stren/designsystem/theme/Color.kt
@@ -1,0 +1,19 @@
+package com.haidoan.android.stren.designsystem.theme
+
+import androidx.compose.ui.graphics.Color
+
+val Red40 = Color(0xFFD31027)
+val Red50 = Color(0xFFEA384D)
+val Red80 = Color(0xFFFFACB6)
+
+val Orange60 = Color(0xFFF37335)
+val Orange80 = Color(0xFFFFB800)
+
+val Gray50 = Color(0xFF7B6F72)
+val Gray60 = Color(0xFFADA4A5)
+val Gray90 = Color(0xFFDDDADA)
+val Gray95 = Color(0xFFEEEEEE)
+
+val Green50 = Color(0xFF3BB44B)
+val Green70 = Color(0xFF4BC85B)
+val Green95 = Color(0xFFBBFFC4)
