+++ app/src/main/java/com/haidoan/android/stren/feat/auth/login/LoginUiState.kt
@@ -1,0 +1,10 @@
+package com.haidoan.android.stren.feat.auth.login
+
+data class LoginUiState(
+    val email: String = "",
+    val password: String = "",
+    val isAuthFailed: Boolean = false,
+    val errorMessage: String = "",
+    val isAuthSuccess: Boolean = false,
+    val isLoading: Boolean = false
+)
