+++ app/src/main/java/com/haidoan/android/stren/feat/auth/utils/ValidationUtils.kt
@@ -1,0 +1,9 @@
+package com.haidoan.android.stren.feat.auth.utils
+
+import android.text.TextUtils
+import android.util.Patterns
+
+internal object ValidationUtils {
+    internal fun isEmailValid(email: String) =
+        !TextUtils.isEmpty(email) && Patterns.EMAIL_ADDRESS.matcher(email).matches();
+}
