+++ app/src/main/java/com/haidoan/android/stren/navigation/Screen.kt
@@ -1,0 +1,6 @@
+package com.haidoan.android.stren.navigation
+
+sealed class Screen(val route: String) {
+    object Login : Screen(route = "login_screen")
+    object Signup : Screen(route = "signup_screen")
+}
